+++ src/main/java/seedu/address/logic/commands/cardcommands/AddCardCommand.java
@@ -1,11 +1,6 @@
 package seedu.address.logic.commands.cardcommands;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
@@ -16,34 +11,29 @@ import seedu.address.model.deck.Deck;
 /**
  * Adds a person to the address book.
  */
-public class CreateCardCommand extends Command {
+public class AddCardCommand extends Command {
 
     public static final String COMMAND_WORD = "add";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a person to the address book. "
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a card to the current deck. "
             + "Parameters: "
-            + PREFIX_NAME + "NAME "
-            + PREFIX_PHONE + "PHONE "
-            + PREFIX_EMAIL + "EMAIL "
-            + PREFIX_ADDRESS + "ADDRESS "
-            + "[" + PREFIX_TAG + "TAG]...\n"
+            + "FRONT "
+            + ": "
+            + "BACK \n"
             + "Example: " + COMMAND_WORD + " "
-            + PREFIX_NAME + "John Doe "
-            + PREFIX_PHONE + "98765432 "
-            + PREFIX_EMAIL + "johnd@example.com "
-            + PREFIX_ADDRESS + "311, Clementi Ave 2, #02-25 "
-            + PREFIX_TAG + "friends "
-            + PREFIX_TAG + "owesMoney";
+            + "ありがとう "
+            + ": "
+            + "thanks";
 
-    public static final String MESSAGE_SUCCESS = "New person added: %1$s";
-    public static final String MESSAGE_DUPLICATE_PERSON = "This person already exists in the address book";
+    public static final String MESSAGE_SUCCESS = "New card added: %1$s";
+    public static final String MESSAGE_DUPLICATE_PERSON = "This card already exists in the deck";
 
     private final Deck toAdd;
 
     /**
      * Creates an AddCommand to add the specified {@code Person}
      */
-    public CreateCardCommand(Deck deck) {
+    public AddCardCommand(Deck deck) {
         requireNonNull(deck);
         toAdd = deck;
     }
@@ -63,7 +53,7 @@ public class CreateCardCommand extends Command {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof CreateCardCommand // instanceof handles nulls
-                && toAdd.equals(((CreateCardCommand) other).toAdd));
+                || (other instanceof AddCardCommand // instanceof handles nulls
+                && toAdd.equals(((AddCardCommand) other).toAdd));
     }
 }
